合同回款

+++ src/main/java/com/wondertek/cpm/web/rest/ContractReceiveResource.java
@@ -1,8 +1,8 @@
 package com.wondertek.cpm.web.rest;
 
-import java.net.URI;
 import java.net.URISyntaxException;
 import java.time.ZonedDateTime;
+import java.util.Date;
 import java.util.List;
 import java.util.Optional;
 
@@ -19,7 +19,6 @@ import org.springframework.security.access.annotation.Secured;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -28,8 +27,8 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.codahale.metrics.annotation.Timed;
 import com.wondertek.cpm.CpmConstants;
+import com.wondertek.cpm.config.DateUtil;
 import com.wondertek.cpm.config.StringUtil;
-import com.wondertek.cpm.domain.ContractReceive;
 import com.wondertek.cpm.domain.ContractReceive;
 import com.wondertek.cpm.domain.vo.ContractReceiveVo;
 import com.wondertek.cpm.security.AuthoritiesConstants;
@@ -52,71 +51,74 @@ public class ContractReceiveResource {
     @Inject
     private ContractReceiveService contractReceiveService;
 
-    @PostMapping("/contract-receives")
-    @Timed
-    @Secured(AuthoritiesConstants.ROLE_CONTRACT_RECEIVE)
-    public ResponseEntity<ContractReceive> createContractReceive(@RequestBody ContractReceive contractReceive) throws URISyntaxException {
-        log.debug("REST request to save ContractReceive : {}", contractReceive);
-        if (contractReceive.getId() != null) {
-            return ResponseEntity.badRequest().headers(HeaderUtil.createFailureAlert("contractReceive", "idexists", "A new contractReceive cannot already have an ID")).body(null);
-        }
-        ContractReceive result = contractReceiveService.save(contractReceive);
-        return ResponseEntity.created(new URI("/api/contract-receives/" + result.getId()))
-            .headers(HeaderUtil.createEntityCreationAlert("contractReceive", result.getId().toString()))
-            .body(result);
-    }
-
     @PutMapping("/contract-receives")
     @Timed
-    public ResponseEntity<Boolean> updateContractReceive(@RequestBody ContractReceive contractReceive) throws URISyntaxException {
+    @Secured(AuthoritiesConstants.ROLE_CONTRACT_RECEIVE)
+    public ResponseEntity<Void> updateContractReceive(@RequestBody ContractReceive contractReceive) throws URISyntaxException {
         log.debug("REST request to update ContractReceive : {}", contractReceive);
         Boolean isNew = contractReceive.getId() == null;
-        if (contractReceive.getContractId() == null || contractReceive.getCreateTime() == null
+        
+        if (contractReceive.getContractId() == null
         		|| StringUtil.isNullStr(contractReceive.getReceiver()) || contractReceive.getReceiveDay() == null
         		|| contractReceive.getReceiveTotal() == null) { 
-        		
-        		return null;
+        	return ResponseEntity.badRequest().headers(HeaderUtil.createError("cpmApp.contractReceive.save.paramNone", "")).body(null);
 		}
+        if (contractReceive.getReceiveTotal() <= 0) { 
+        	return ResponseEntity.badRequest().headers(HeaderUtil.createError("cpmApp.contractReceive.save.receiveTotalError", "")).body(null);
+		}
+        Long today = StringUtil.nullToLong(DateUtil.formatDate(DateUtil.DATE_YYYYMMDD_PATTERN, new Date()));
+        if(contractReceive.getReceiveDay() > today){
+        	return ResponseEntity.badRequest().headers(HeaderUtil.createError("cpmApp.contractReceive.save.receiveDayError", "")).body(null);
+        }
         String updator = SecurityUtils.getCurrentUserLogin();
         ZonedDateTime updateTime = ZonedDateTime.now();
+        Double oldTotal = 0d;
         if (isNew) {
 			contractReceive.setStatus(CpmConstants.STATUS_VALID);
 			contractReceive.setCreateTime(updateTime);
+			contractReceive.setCreator(updator);
 		}else {
 			ContractReceiveVo contractReceiveVo = contractReceiveService.getContractReceive(contractReceive.getId());
 			if (contractReceiveVo == null) {
-				return null;
+				return ResponseEntity.badRequest().headers(HeaderUtil.createError("cpmApp.contractReceive.save.noPerm", "")).body(null);
 			}
-			ContractReceive old = contractReceiveService.findOne(contractReceive.getId());
-			if(old == null){
-        		return ResponseEntity.badRequest().headers(HeaderUtil.createError("cpmApp.projectCost.save.idNone", "")).body(null);
-        	}else if(old.getContractId() != contractReceive.getContractId().longValue()){
-        		return null;
-        	}else if(old.getStatus() == CpmConstants.STATUS_DELETED){
-        		return null;
+			if(contractReceiveVo.getContractId() != contractReceive.getContractId().longValue()){
+        		return ResponseEntity.badRequest().headers(HeaderUtil.createError("cpmApp.contractReceive.save.contractChanged", "")).body(null);
+        	}else if(contractReceiveVo.getStatus() == CpmConstants.STATUS_DELETED){
+        		return ResponseEntity.badRequest().headers(HeaderUtil.createError("cpmApp.contractReceive.save.status2Error", "")).body(null);
         	}
-			contractReceive.setStatus(old.getStatus());
-			contractReceive.setCreateTime(old.getCreateTime());
-			contractReceive.setCreator(old.getCreator());
+			oldTotal = contractReceiveVo.getReceiveTotal();
+			contractReceive.setStatus(contractReceiveVo.getStatus());
+			contractReceive.setCreateTime(contractReceiveVo.getCreateTime());
+			contractReceive.setCreator(contractReceiveVo.getCreator());
 		}
         contractReceive.setUpdateTime(updateTime);
         contractReceive.setUpdator(updator);
-        ContractReceive result = contractReceiveService.save(contractReceive);
-        return ResponseEntity.ok()
-            .headers(HeaderUtil.createEntityUpdateAlert("contractReceive", result.getId().toString()))
-            .body(isNew);
+        ContractReceive result = contractReceiveService.save(contractReceive,oldTotal);
+        
+        if(isNew){
+        	return ResponseEntity.ok()
+        			.headers(HeaderUtil.createEntityCreationAlert("contractReceive", result.getId().toString()))
+        			.body(null);
+        }else{
+        	return ResponseEntity.ok()
+        			.headers(HeaderUtil.createEntityUpdateAlert("contractReceive", result.getId().toString()))
+        			.body(null);
+        }
     }
 
     @GetMapping("/contract-receives")
     @Timed
-    public ResponseEntity<List<ContractReceiveVo>> getAllContractReceives(@RequestParam(value="contractId",required = false) Long contractId,@ApiParam Pageable pageable)
+    @Secured(AuthoritiesConstants.ROLE_CONTRACT_RECEIVE)
+    public ResponseEntity<List<ContractReceiveVo>> getAllContractReceives(
+    		@RequestParam(value="contractId",required = false) Long contractId,
+    		@ApiParam Pageable pageable)
         throws URISyntaxException {
         log.debug("REST request to get a page of ContractReceives");
         ContractReceive contractReceive = new ContractReceive();
         contractReceive.setContractId(contractId);
         
-        
-        Page<ContractReceiveVo> page = contractReceiveService.getuserPage(contractReceive,pageable);
+        Page<ContractReceiveVo> page = contractReceiveService.getUserPage(contractReceive,pageable);
         
         HttpHeaders headers = PaginationUtil.generatePaginationHttpHeaders(page, "/api/contract-receives");
         return new ResponseEntity<>(page.getContent(), headers, HttpStatus.OK);
@@ -140,6 +142,13 @@ public class ContractReceiveResource {
     @Secured(AuthoritiesConstants.ROLE_CONTRACT_RECEIVE)
     public ResponseEntity<Void> deleteContractReceive(@PathVariable Long id) {
         log.debug("REST request to delete ContractReceive : {}", id);
+        ContractReceiveVo contractReceiveVo = contractReceiveService.getContractReceive(id);
+        if(contractReceiveVo == null){
+        	return ResponseEntity.badRequest().headers(HeaderUtil.createError("cpmApp.contractReceive.save.noPerm", "")).body(null);
+        }
+        if(contractReceiveVo.getStatus() == CpmConstants.STATUS_DELETED){
+        	return ResponseEntity.badRequest().headers(HeaderUtil.createError("cpmApp.contractReceive.delete.status2Error", "")).body(null);
+        }
         contractReceiveService.delete(id);
         return ResponseEntity.ok().headers(HeaderUtil.createEntityDeletionAlert("contractReceive", id.toString())).build();
     }
